requirements.txt
scikit-learn 
pandas 
pyyaml 
matplotlib 
tensorflow
